requirements.txt
click_shell
click
gtrending
gitpython